updated the feeling loopy loop with an if statement and nested if statement.
I was overthinking the logic here. I thought I needed a nested loop for the csv portion. But was given more perspective on the logic and what to look out for. Instead of disregarding commas in the string, I should have been keeping track of them with a counter variable. That way I can flag my if statements to look for the counter variable and assign whatever was logged into the cell.